Remove code that's now unused
