+++ pkg/http/universalapi.go
@@ -15,12 +15,10 @@ package http
 
 import (
 	"context"
-	"encoding/json"
 	"io"
 	"net/http"
 	"reflect"
 
-	"github.com/valyala/fasthttp"
 	"google.golang.org/protobuf/encoding/protojson"
 	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/reflect/protoreflect"
@@ -28,14 +26,8 @@ import (
 	"github.com/dapr/dapr/pkg/messages"
 )
 
-// Object containing options for the UniversalFastHTTPHandler method.
+// Object containing options for the UniversalHTTPHandler method.
 type UniversalHTTPHandlerOpts[T proto.Message, U proto.Message] struct {
-	// This modifier allows modifying the input proto object before the handler is called. This property is optional.
-	// The input proto object contantains all properties parsed from the request's body (for non-GET requests), and this modifier can alter it for example with properties from the URL (to make APIs RESTful).
-	// The modifier should return the modified object.
-	// TODO: Remove this when FastHTTP support is dropped.
-	InModifierFastHTTP func(reqCtx *fasthttp.RequestCtx, in T) (T, error)
-
 	// This modifier allows modifying the input proto object before the handler is called. This property is optional.
 	// The input proto object contantains all properties parsed from the request's body (for non-GET requests), and this modifier can alter it for example with properties from the URL (to make APIs RESTful).
 	// The modifier should return the modified object.
@@ -59,7 +51,7 @@ type UniversalHTTPHandlerOpts[T proto.Message, U proto.Message] struct {
 	ProtoResponseEmitUnpopulated bool
 }
 
-// UniversalFastHTTPHandler wraps a Universal API method into a HTTP handler.
+// UniversalHTTPHandler wraps a Universal API method into a HTTP handler.
 func UniversalHTTPHandler[T proto.Message, U proto.Message](
 	handler func(ctx context.Context, in T) (U, error),
 	opts UniversalHTTPHandlerOpts[T, U],
@@ -158,101 +150,6 @@ func UniversalHTTPHandler[T proto.Message, U proto.Message](
 	}
 }
 
-// UniversalFastHTTPHandler wraps a UniversalAPI method into a FastHTTP handler.
-func UniversalFastHTTPHandler[T proto.Message, U proto.Message](
-	handler func(ctx context.Context, in T) (U, error),
-	opts UniversalHTTPHandlerOpts[T, U],
-) fasthttp.RequestHandler {
-	var zero T
-	rt := reflect.ValueOf(zero).Type().Elem()
-	pjsonDec := protojson.UnmarshalOptions{
-		DiscardUnknown: true,
-		AllowPartial:   true,
-	}
-
-	return func(reqCtx *fasthttp.RequestCtx) {
-		var err error
-
-		// Need to use some reflection magic to allocate a value for the pointer of the generic type T
-		in := reflect.New(rt).Interface().(T)
-
-		// Parse the body as JSON
-		if !opts.SkipInputBody {
-			// Read the request body and decode it as JSON using protojson
-			body := reqCtx.PostBody()
-			if len(body) > 0 {
-				err = pjsonDec.Unmarshal(body, in)
-				if err != nil {
-					msg := messages.ErrMalformedRequest.WithFormat(err)
-					universalFastHTTPErrorResponder(reqCtx, msg)
-					log.Debug(msg)
-					return
-				}
-			}
-		}
-
-		// If we have an inModifier function, invoke it now
-		if opts.InModifierFastHTTP != nil {
-			in, err = opts.InModifierFastHTTP(reqCtx, in)
-			if err != nil {
-				universalFastHTTPErrorResponder(reqCtx, err)
-				return
-			}
-		}
-
-		// Invoke the gRPC handler
-		// We need to create a context specific for this because fasthttp's reqCtx is tied to the server's lifecycle and not the request's
-		// See: https://github.com/valyala/fasthttp/issues/1350
-		ctx, cancel := context.WithCancel(reqCtx)
-		res, err := handler(ctx, in)
-		cancel()
-		if err != nil {
-			// Error is already logged by the handlers, we won't log it again
-			universalFastHTTPErrorResponder(reqCtx, err)
-			return
-		}
-
-		if reflect.ValueOf(res).IsZero() {
-			fasthttpRespond(reqCtx, fasthttpResponseWithEmpty())
-			return
-		}
-
-		// Set success status code to 200 if none is specified
-		if opts.SuccessStatusCode == 0 {
-			opts.SuccessStatusCode = http.StatusOK
-		}
-
-		// If we do not have an output modifier, respond right away
-		if opts.OutModifier == nil {
-			universalFastHTTPProtoResponder(reqCtx, res, opts.SuccessStatusCode, opts.ProtoResponseEmitUnpopulated)
-			return
-		}
-
-		// Invoke the modifier
-		newRes, err := opts.OutModifier(res)
-		if err != nil {
-			universalFastHTTPErrorResponder(reqCtx, err)
-			return
-		}
-
-		// Check if the response is a proto object (which is encoded with protojson) or any other object to encode as regular JSON
-		switch m := newRes.(type) {
-		case nil:
-			fasthttpRespond(reqCtx, fasthttpResponseWithEmpty())
-			return
-		case *UniversalHTTPRawResponse:
-			universalFastHTTPRawResponder(reqCtx, m, opts.SuccessStatusCode)
-			return
-		case protoreflect.ProtoMessage:
-			universalFastHTTPProtoResponder(reqCtx, m, opts.SuccessStatusCode, opts.ProtoResponseEmitUnpopulated)
-			return
-		default:
-			universalFastHTTPJSONResponder(reqCtx, m, opts.SuccessStatusCode)
-			return
-		}
-	}
-}
-
 // Contains a pre-serialized response as well as its content type.
 // An OutModifier can return this object if it needs to serialize the response itself.
 type UniversalHTTPRawResponse struct {
@@ -262,60 +159,4 @@ type UniversalHTTPRawResponse struct {
 	ContentType string
 	// Optional status code; if empty, uses the default SuccessStatusCode.
 	StatusCode int
-}
-
-func universalFastHTTPRawResponder(reqCtx *fasthttp.RequestCtx, m *UniversalHTTPRawResponse, statusCode int) {
-	if m.StatusCode > 0 {
-		statusCode = m.StatusCode
-	}
-	if m.ContentType != "" {
-		reqCtx.Response.Header.SetContentType(m.ContentType)
-	}
-
-	fasthttpRespond(reqCtx, fasthttpResponseWith(statusCode, m.Body))
-}
-
-func universalFastHTTPProtoResponder(reqCtx *fasthttp.RequestCtx, m protoreflect.ProtoMessage, statusCode int, emitUnpopulated bool) {
-	// Encode the response as JSON using protojson
-	respBytes, err := protojson.MarshalOptions{
-		EmitUnpopulated: emitUnpopulated,
-	}.Marshal(m)
-	if err != nil {
-		msg := NewErrorResponse("ERR_INTERNAL", "failed to encode response as JSON: "+err.Error())
-		fasthttpRespond(reqCtx, fasthttpResponseWithError(http.StatusInternalServerError, msg))
-		log.Debug(msg)
-		return
-	}
-
-	fasthttpRespond(reqCtx, fasthttpResponseWithJSON(statusCode, respBytes))
-}
-
-func universalFastHTTPJSONResponder(reqCtx *fasthttp.RequestCtx, m any, statusCode int) {
-	// Encode the response as JSON using the regular JSON package
-	respBytes, err := json.Marshal(m)
-	if err != nil {
-		msg := NewErrorResponse("ERR_INTERNAL", "failed to encode response as JSON: "+err.Error())
-		fasthttpRespond(reqCtx, fasthttpResponseWithError(http.StatusInternalServerError, msg))
-		log.Debug(msg)
-		return
-	}
-
-	fasthttpRespond(reqCtx, fasthttpResponseWithJSON(statusCode, respBytes))
-}
-
-func universalFastHTTPErrorResponder(reqCtx *fasthttp.RequestCtx, err error) {
-	if err == nil {
-		return
-	}
-
-	// Check if it's an APIError object
-	apiErr, ok := err.(messages.APIError)
-	if ok {
-		fasthttpRespond(reqCtx, fasthttpResponseWithError(apiErr.HTTPCode(), apiErr))
-		return
-	}
-
-	// Respond with a generic error
-	msg := NewErrorResponse("ERROR", err.Error())
-	fasthttpRespond(reqCtx, fasthttpResponseWithError(http.StatusInternalServerError, msg))
 }
